gates, scaling, and enemy coding
The gates are in with their animations and are already set to open when all the enemies in each level are killed. I also started programming the enemy script but very little. I fixed the material scaling for most of the walls and i made the spiders sizes all match.